# Harden ref-counting concurrency semantics