--- a/buffer/src/test/java/io/netty/buffer/AbstractReferenceCountedByteBufTest.java
+++ b/buffer/src/test/java/io/netty/buffer/AbstractReferenceCountedByteBufTest.java
@@ -38,14 +38,16 @@ public class AbstractReferenceCountedByteBufTest {
         AbstractReferenceCountedByteBuf referenceCounted = newReferenceCounted();
         referenceCounted.setRefCnt(Integer.MAX_VALUE);
         assertEquals(Integer.MAX_VALUE, referenceCounted.refCnt());
-        referenceCounted.retain();
+        // extra retain since we can actually go one higher before overflowing
+        referenceCounted.retain().retain();
     }
 
     @Test(expected = IllegalReferenceCountException.class)
     public void testRetainOverflow2() {
         AbstractReferenceCountedByteBuf referenceCounted = newReferenceCounted();
         assertEquals(1, referenceCounted.refCnt());
-        referenceCounted.retain(Integer.MAX_VALUE);
+        // extra retain since we can actually go one higher before overflowing
+        referenceCounted.retain().retain(Integer.MAX_VALUE);
     }
 
     @Test(expected = IllegalReferenceCountException.class)