--- a/common/src/main/java/io/netty/util/AbstractReferenceCounted.java
+++ b/common/src/main/java/io/netty/util/AbstractReferenceCounted.java
@@ -15,30 +15,58 @@
  */
 package io.netty.util;
 
+import static io.netty.util.internal.ObjectUtil.checkPositive;
+
 import java.util.concurrent.atomic.AtomicIntegerFieldUpdater;
 
-import static io.netty.util.internal.ObjectUtil.checkPositive;
+import io.netty.util.internal.PlatformDependent;
 
 /**
  * Abstract base class for classes wants to implement {@link ReferenceCounted}.
  */
 public abstract class AbstractReferenceCounted implements ReferenceCounted {
-
+    private static final long REFCNT_FIELD_OFFSET;
     private static final AtomicIntegerFieldUpdater<AbstractReferenceCounted> refCntUpdater =
             AtomicIntegerFieldUpdater.newUpdater(AbstractReferenceCounted.class, "refCnt");
 
-    private volatile int refCnt = 1;
+    // even => "real" refcount is (refCnt >>> 1); odd => "real" refcount is 0
+    @SuppressWarnings("unused")
+    private volatile int refCnt = 2;
+
+    static {
+        long refCntFieldOffset = -1;
+        try {
+            if (PlatformDependent.hasUnsafe()) {
+                refCntFieldOffset = PlatformDependent.objectFieldOffset(
+                        AbstractReferenceCounted.class.getDeclaredField("refCnt"));
+            }
+        } catch (Throwable ignore) {
+            refCntFieldOffset = -1;
+        }
+
+        REFCNT_FIELD_OFFSET = refCntFieldOffset;
+    }
+
+    private static int realRef(int refCnt) {
+        return (refCnt & 1) != 0 ? 0 : refCnt >>> 1;
+    }
+
+    private int nonVolatileRawCnt() {
+        // TODO: Once we compile against later versions of Java we can replace the Unsafe usage here by varhandles.
+        return REFCNT_FIELD_OFFSET != -1 ? PlatformDependent.getInt(this, REFCNT_FIELD_OFFSET)
+                : refCntUpdater.get(this);
+    }
 
     @Override
-    public final int refCnt() {
-        return refCnt;
+    public int refCnt() {
+        return realRef(refCntUpdater.get(this));
     }
 
     /**
      * An unsafe operation intended for use by a subclass that sets the reference count of the buffer directly
      */
     protected final void setRefCnt(int refCnt) {
-        refCntUpdater.set(this, refCnt);
+        refCntUpdater.set(this, refCnt << 1); // overflow OK here
     }
 
     @Override
@@ -51,12 +79,19 @@ public abstract class AbstractReferenceCounted implements ReferenceCounted {
         return retain0(checkPositive(increment, "increment"));
     }
 
-    private ReferenceCounted retain0(int increment) {
-        int oldRef = refCntUpdater.getAndAdd(this, increment);
-        if (oldRef <= 0 || oldRef + increment < oldRef) {
-            // Ensure we don't resurrect (which means the refCnt was 0) and also that we encountered an overflow.
+    private ReferenceCounted retain0(final int increment) {
+        // all changes to the raw count are 2x the "real" change
+        int adjustIncrement = increment << 1; // overflow OK here
+        int oldRef = refCntUpdater.getAndAdd(this, adjustIncrement);
+        if ((oldRef & 1) != 0) {
+            throw new IllegalReferenceCountException(0, increment);
+        }
+        // don't pass 0!
+        if ((oldRef <= 0 && oldRef + adjustIncrement >= 0)
+                || (oldRef >= 0 && oldRef + adjustIncrement < oldRef)) {
+            // overflow case
             refCntUpdater.getAndAdd(this, -increment);
-            throw new IllegalReferenceCountException(oldRef, increment);
+            throw new IllegalReferenceCountException(realRef(oldRef), increment);
         }
         return this;
     }
@@ -77,16 +112,36 @@ public abstract class AbstractReferenceCounted implements ReferenceCounted {
     }
 
     private boolean release0(int decrement) {
-        int oldRef = refCntUpdater.getAndAdd(this, -decrement);
-        if (oldRef == decrement) {
-            deallocate();
-            return true;
-        } else if (oldRef < decrement || oldRef - decrement > oldRef) {
-            // Ensure we don't over-release, and avoid underflow.
-            refCntUpdater.getAndAdd(this, decrement);
-            throw new IllegalReferenceCountException(oldRef, -decrement);
+        int rawCnt = nonVolatileRawCnt();
+        for (boolean firstTry = true;; firstTry = false) {
+            if ((rawCnt & 1) != 0) {
+                throw new IllegalReferenceCountException(0, -decrement);
+            }
+            int realCnt = rawCnt >>> 1;
+            if (decrement == realCnt) {
+                // most likely case
+                if (refCntUpdater.compareAndSet(this, rawCnt, 1)) { // any odd number
+                    deallocate();
+                    return true;
+                }
+                if (!firstTry) {
+                    // this benefits throughput under high contention
+                    Thread.yield();
+                }
+            } else if (decrement < realCnt) {
+                // all changes to the raw count are 2x the "real" change
+                if (refCntUpdater.compareAndSet(this, rawCnt, rawCnt - (decrement << 1))) {
+                    return false;
+                }
+                if (!firstTry) {
+                    // this benefits throughput under high contention
+                    Thread.yield();
+                }
+            } else if (!firstTry) {
+                throw new IllegalReferenceCountException(realCnt, -decrement);
+            }
+            rawCnt = refCntUpdater.get(this);
         }
-        return false;
     }
 
     /**