--- a/common/src/test/java/io/netty/util/AbstractReferenceCountedTest.java
+++ b/common/src/test/java/io/netty/util/AbstractReferenceCountedTest.java
@@ -19,16 +19,12 @@ import io.netty.util.internal.ThreadLocalRandom;
 import org.junit.Test;
 
 import java.util.ArrayDeque;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Queue;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicReference;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;