@normanmaurer 2018-11-20
@netty-bot test this please