@njhill 2018-11-20
Benchmark results:

Pre- 83a19d565064ee36998eb94f946e5a4264001065:
```
Benchmark                           (delay)    Mode       Cnt         Score    Error  Units
retainReleaseContended                    1  sample  36460170      1152.856 ± 14.343  ns/op
retainReleaseContended·p0.00              1  sample                  37.000           ns/op
retainReleaseContended·p0.50              1  sample                 807.000           ns/op
retainReleaseContended·p0.90              1  sample                1858.000           ns/op
retainReleaseContended·p0.95              1  sample                2320.000           ns/op
retainReleaseContended·p0.99              1  sample                3524.000           ns/op
retainReleaseContended·p0.999             1  sample               16704.000           ns/op
retainReleaseContended·p0.9999            1  sample              250880.000           ns/op
retainReleaseContended·p1.00              1  sample            32309248.000           ns/op
retainReleaseContended                   10  sample  38184690      1119.333 ± 15.226  ns/op
retainReleaseContended·p0.00             10  sample                  45.000           ns/op
retainReleaseContended·p0.50             10  sample                 770.000           ns/op
retainReleaseContended·p0.90             10  sample                1748.000           ns/op
retainReleaseContended·p0.95             10  sample                2184.000           ns/op
retainReleaseContended·p0.99             10  sample                3312.000           ns/op
retainReleaseContended·p0.999            10  sample               16224.000           ns/op
retainReleaseContended·p0.9999           10  sample              258695.910           ns/op
retainReleaseContended·p1.00             10  sample            28803072.000           ns/op
retainReleaseContended                  100  sample  40444975      1040.120 ± 11.997  ns/op
retainReleaseContended·p0.00            100  sample                 229.000           ns/op
retainReleaseContended·p0.50            100  sample                 791.000           ns/op
retainReleaseContended·p0.90            100  sample                1482.000           ns/op
retainReleaseContended·p0.95            100  sample                1780.000           ns/op
retainReleaseContended·p0.99            100  sample                2540.000           ns/op
retainReleaseContended·p0.999           100  sample               14528.000           ns/op
retainReleaseContended·p0.9999          100  sample              184576.000           ns/op
retainReleaseContended·p1.00            100  sample            27295744.000           ns/op
retainReleaseContended                 1000  sample  29429689      2842.698 ± 26.030  ns/op
retainReleaseContended·p0.00           1000  sample                2164.000           ns/op
retainReleaseContended·p0.50           1000  sample                2444.000           ns/op
retainReleaseContended·p0.90           1000  sample                2520.000           ns/op
retainReleaseContended·p0.95           1000  sample                2588.000           ns/op
retainReleaseContended·p0.99           1000  sample                2820.000           ns/op
retainReleaseContended·p0.999          1000  sample               38912.000           ns/op
retainReleaseContended·p0.9999         1000  sample              535552.000           ns/op
retainReleaseContended·p1.00           1000  sample            29949952.000           ns/op
retainReleaseContended                10000  sample  24367466     25048.634 ± 56.891  ns/op
retainReleaseContended·p0.00          10000  sample               20736.000           ns/op
retainReleaseContended·p0.50          10000  sample               23616.000           ns/op
retainReleaseContended·p0.90          10000  sample               23744.000           ns/op
retainReleaseContended·p0.95          10000  sample               24352.000           ns/op
retainReleaseContended·p0.99          10000  sample               38144.000           ns/op
retainReleaseContended·p0.999         10000  sample              273680.896           ns/op
retainReleaseContended·p0.9999        10000  sample             2613248.000           ns/op
retainReleaseContended·p1.00          10000  sample            59965440.000           ns/op
retainReleaseUncontended                1  sample   7509511        55.367 ±  2.721  ns/op
retainReleaseUncontended·p0.00          1  sample                  35.000           ns/op
retainReleaseUncontended·p0.50          1  sample                  41.000           ns/op
retainReleaseUncontended·p0.90          1  sample                  52.000           ns/op
retainReleaseUncontended·p0.95          1  sample                  62.000           ns/op
retainReleaseUncontended·p0.99          1  sample                  98.000           ns/op
retainReleaseUncontended·p0.999         1  sample                 359.488           ns/op
retainReleaseUncontended·p0.9999        1  sample               18305.562           ns/op
retainReleaseUncontended·p1.00          1  sample             5824512.000           ns/op
retainReleaseUncontended               10  sample   6129780        73.046 ±  5.091  ns/op
retainReleaseUncontended·p0.00         10  sample                  43.000           ns/op
retainReleaseUncontended·p0.50         10  sample                  53.000           ns/op
retainReleaseUncontended·p0.90         10  sample                  68.000           ns/op
retainReleaseUncontended·p0.95         10  sample                  73.000           ns/op
retainReleaseUncontended·p0.99         10  sample                 106.000           ns/op
retainReleaseUncontended·p0.999        10  sample                2280.876           ns/op
retainReleaseUncontended·p0.9999       10  sample               18816.000           ns/op
retainReleaseUncontended·p1.00         10  sample             8978432.000           ns/op
retainReleaseUncontended              100  sample   6601891       279.857 ±  1.515  ns/op
retainReleaseUncontended·p0.00        100  sample                 219.000           ns/op
retainReleaseUncontended·p0.50        100  sample                 247.000           ns/op
retainReleaseUncontended·p0.90        100  sample                 273.000           ns/op
retainReleaseUncontended·p0.95        100  sample                 291.000           ns/op
retainReleaseUncontended·p0.99        100  sample                 337.000           ns/op
retainReleaseUncontended·p0.999       100  sample               13824.000           ns/op
retainReleaseUncontended·p0.9999      100  sample               21440.000           ns/op
retainReleaseUncontended·p1.00        100  sample             2428928.000           ns/op
retainReleaseUncontended             1000  sample   5385984      2354.879 ±  6.543  ns/op
retainReleaseUncontended·p0.00       1000  sample                1970.000           ns/op
retainReleaseUncontended·p0.50       1000  sample                2156.000           ns/op
retainReleaseUncontended·p0.90       1000  sample                2312.000           ns/op
retainReleaseUncontended·p0.95       1000  sample                2448.000           ns/op
retainReleaseUncontended·p0.99       1000  sample               15088.000           ns/op
retainReleaseUncontended·p0.999      1000  sample               20896.000           ns/op
retainReleaseUncontended·p0.9999     1000  sample               32793.696           ns/op
retainReleaseUncontended·p1.00       1000  sample             4546560.000           ns/op
retainReleaseUncontended            10000  sample   4314810     23143.254 ± 28.804  ns/op
retainReleaseUncontended·p0.00      10000  sample               19488.000           ns/op
retainReleaseUncontended·p0.50      10000  sample               21344.000           ns/op
retainReleaseUncontended·p0.90      10000  sample               26144.000           ns/op
retainReleaseUncontended·p0.95      10000  sample               35712.000           ns/op
retainReleaseUncontended·p0.99      10000  sample               50688.000           ns/op
retainReleaseUncontended·p0.999     10000  sample               65792.000           ns/op
retainReleaseUncontended·p0.9999    10000  sample              143492.838           ns/op
retainReleaseUncontended·p1.00      10000  sample            12255232.000           ns/op
```

Current (before):
```
Benchmark                           (delay)    Mode       Cnt         Score    Error  Units
retainReleaseContended                    1  sample  33617682       396.846 ± 12.371  ns/op
retainReleaseContended·p0.00              1  sample                  36.000           ns/op
retainReleaseContended·p0.50              1  sample                 267.000           ns/op
retainReleaseContended·p0.90              1  sample                 484.000           ns/op
retainReleaseContended·p0.95              1  sample                 558.000           ns/op
retainReleaseContended·p0.99              1  sample                 725.000           ns/op
retainReleaseContended·p0.999             1  sample                5232.000           ns/op
retainReleaseContended·p0.9999            1  sample              124544.000           ns/op
retainReleaseContended·p1.00              1  sample            40239104.000           ns/op
retainReleaseContended                   10  sample  30765523       422.296 ± 12.817  ns/op
retainReleaseContended·p0.00             10  sample                  45.000           ns/op
retainReleaseContended·p0.50             10  sample                 289.000           ns/op
retainReleaseContended·p0.90             10  sample                 479.000           ns/op
retainReleaseContended·p0.95             10  sample                 551.000           ns/op
retainReleaseContended·p0.99             10  sample                 728.000           ns/op
retainReleaseContended·p0.999            10  sample                5888.000           ns/op
retainReleaseContended·p0.9999           10  sample              140032.000           ns/op
retainReleaseContended·p1.00             10  sample            23265280.000           ns/op
retainReleaseContended                  100  sample  26334520       475.976 ± 13.347  ns/op
retainReleaseContended·p0.00            100  sample                 235.000           ns/op
retainReleaseContended·p0.50            100  sample                 353.000           ns/op
retainReleaseContended·p0.90            100  sample                 433.000           ns/op
retainReleaseContended·p0.95            100  sample                 462.000           ns/op
retainReleaseContended·p0.99            100  sample                 535.000           ns/op
retainReleaseContended·p0.999           100  sample               11152.000           ns/op
retainReleaseContended·p0.9999          100  sample              156300.262           ns/op
retainReleaseContended·p1.00            100  sample            26771456.000           ns/op
retainReleaseContended                 1000  sample  29409220      2954.896 ± 35.426  ns/op
retainReleaseContended·p0.00           1000  sample                2096.000           ns/op
retainReleaseContended·p0.50           1000  sample                2424.000           ns/op
retainReleaseContended·p0.90           1000  sample                2468.000           ns/op
retainReleaseContended·p0.95           1000  sample                2496.000           ns/op
retainReleaseContended·p0.99           1000  sample                2668.000           ns/op
retainReleaseContended·p0.999          1000  sample               56192.000           ns/op
retainReleaseContended·p0.9999         1000  sample              668672.000           ns/op
retainReleaseContended·p1.00           1000  sample            49283072.000           ns/op
retainReleaseContended                10000  sample  24335420     25097.878 ± 65.738  ns/op
retainReleaseContended·p0.00          10000  sample               21344.000           ns/op
retainReleaseContended·p0.50          10000  sample               23616.000           ns/op
retainReleaseContended·p0.90          10000  sample               23712.000           ns/op
retainReleaseContended·p0.95          10000  sample               24352.000           ns/op
retainReleaseContended·p0.99          10000  sample               38976.000           ns/op
retainReleaseContended·p0.999         10000  sample              253952.000           ns/op
retainReleaseContended·p0.9999        10000  sample             2793472.000           ns/op
retainReleaseContended·p1.00          10000  sample            58261504.000           ns/op
retainReleaseUncontended                1  sample   4484558        58.703 ±  1.546  ns/op
retainReleaseUncontended·p0.00          1  sample                  33.000           ns/op
retainReleaseUncontended·p0.50          1  sample                  39.000           ns/op
retainReleaseUncontended·p0.90          1  sample                  50.000           ns/op
retainReleaseUncontended·p0.95          1  sample                  65.000           ns/op
retainReleaseUncontended·p0.99          1  sample                  98.000           ns/op
retainReleaseUncontended·p0.999         1  sample                4296.000           ns/op
retainReleaseUncontended·p0.9999        1  sample               19889.411           ns/op
retainReleaseUncontended·p1.00          1  sample             1775616.000           ns/op
retainReleaseUncontended               10  sample   6199062        73.845 ±  1.339  ns/op
retainReleaseUncontended·p0.00         10  sample                  43.000           ns/op
retainReleaseUncontended·p0.50         10  sample                  53.000           ns/op
retainReleaseUncontended·p0.90         10  sample                  71.000           ns/op
retainReleaseUncontended·p0.95         10  sample                  87.000           ns/op
retainReleaseUncontended·p0.99         10  sample                 122.000           ns/op
retainReleaseUncontended·p0.999        10  sample                2643.748           ns/op
retainReleaseUncontended·p0.9999       10  sample               18752.000           ns/op
retainReleaseUncontended·p1.00         10  sample             1251328.000           ns/op
retainReleaseUncontended              100  sample   6570441       283.853 ±  5.069  ns/op
retainReleaseUncontended·p0.00        100  sample                 219.000           ns/op
retainReleaseUncontended·p0.50        100  sample                 244.000           ns/op
retainReleaseUncontended·p0.90        100  sample                 268.000           ns/op
retainReleaseUncontended·p0.95        100  sample                 280.000           ns/op
retainReleaseUncontended·p0.99        100  sample                 325.000           ns/op
retainReleaseUncontended·p0.999       100  sample               13952.000           ns/op
retainReleaseUncontended·p0.9999      100  sample               21856.000           ns/op
retainReleaseUncontended·p1.00        100  sample             9158656.000           ns/op
retainReleaseUncontended             1000  sample   5345561      2368.951 ±  4.646  ns/op
retainReleaseUncontended·p0.00       1000  sample                1970.000           ns/op
retainReleaseUncontended·p0.50       1000  sample                2112.000           ns/op
retainReleaseUncontended·p0.90       1000  sample                2268.000           ns/op
retainReleaseUncontended·p0.95       1000  sample                2376.000           ns/op
retainReleaseUncontended·p0.99       1000  sample               15424.000           ns/op
retainReleaseUncontended·p0.999      1000  sample               21184.000           ns/op
retainReleaseUncontended·p0.9999     1000  sample               30862.202           ns/op
retainReleaseUncontended·p1.00       1000  sample             2605056.000           ns/op
retainReleaseUncontended            10000  sample   4293861     23226.581 ± 20.925  ns/op
retainReleaseUncontended·p0.00      10000  sample               19488.000           ns/op
retainReleaseUncontended·p0.50      10000  sample               21344.000           ns/op
retainReleaseUncontended·p0.90      10000  sample               27968.000           ns/op
retainReleaseUncontended·p0.95      10000  sample               35840.000           ns/op
retainReleaseUncontended·p0.99      10000  sample               51776.000           ns/op
retainReleaseUncontended·p0.999     10000  sample               65408.000           ns/op
retainReleaseUncontended·p0.9999    10000  sample               98048.000           ns/op
retainReleaseUncontended·p1.00      10000  sample            12369920.000           ns/op
```

This PR (after):
```
Benchmark                           (delay)    Mode       Cnt         Score    Error  Units
retainReleaseContended                    1  sample  30710079       566.717 ± 13.469  ns/op
retainReleaseContended·p0.00              1  sample                  40.000           ns/op
retainReleaseContended·p0.50              1  sample                  82.000           ns/op
retainReleaseContended·p0.90              1  sample                 337.000           ns/op
retainReleaseContended·p0.95              1  sample                1972.000           ns/op
retainReleaseContended·p0.99              1  sample                8528.000           ns/op
retainReleaseContended·p0.999             1  sample               27488.000           ns/op
retainReleaseContended·p0.9999            1  sample              211200.000           ns/op
retainReleaseContended·p1.00              1  sample            40632320.000           ns/op
retainReleaseContended                   10  sample  33563252       584.657 ± 14.439  ns/op
retainReleaseContended·p0.00             10  sample                  48.000           ns/op
retainReleaseContended·p0.50             10  sample                  69.000           ns/op
retainReleaseContended·p0.90             10  sample                 223.000           ns/op
retainReleaseContended·p0.95             10  sample                1386.000           ns/op
retainReleaseContended·p0.99             10  sample               10288.000           ns/op
retainReleaseContended·p0.999            10  sample               39872.000           ns/op
retainReleaseContended·p0.9999           10  sample              217088.000           ns/op
retainReleaseContended·p1.00             10  sample            42663936.000           ns/op
retainReleaseContended                  100  sample  26297644       948.177 ± 18.478  ns/op
retainReleaseContended·p0.00            100  sample                 222.000           ns/op
retainReleaseContended·p0.50            100  sample                 386.000           ns/op
retainReleaseContended·p0.90            100  sample                1666.000           ns/op
retainReleaseContended·p0.95            100  sample                2928.000           ns/op
retainReleaseContended·p0.99            100  sample                6704.000           ns/op
retainReleaseContended·p0.999           100  sample               29472.000           ns/op
retainReleaseContended·p0.9999          100  sample              319488.000           ns/op
retainReleaseContended·p1.00            100  sample            55312384.000           ns/op
retainReleaseContended                 1000  sample  29410837      2857.274 ± 24.380  ns/op
retainReleaseContended·p0.00           1000  sample                1984.000           ns/op
retainReleaseContended·p0.50           1000  sample                2440.000           ns/op
retainReleaseContended·p0.90           1000  sample                2500.000           ns/op
retainReleaseContended·p0.95           1000  sample                2552.000           ns/op
…